@@ -3,15 +3,6 @@
 import { SWRConfig } from "swr";
 import { useState } from "react";
 
-
-const fetcher = async (url: string) => {
-  const res = await fetch(url);
-  if (!res.ok) {
-    throw new Error(`API error: ${res.status}`);
-  }
-  return res.json();
-};
-
 export default function Providers({ children }: { children: React.ReactNode }) {
   const [provider] = useState(() => new Map());
 
@@ -19,14 +10,12 @@ export default function Providers({ children }: { children: React.ReactNode }) {
     <SWRConfig
       value={{
         provider: () => provider,
-        fetcher, 
-        revalidateOnFocus: false, 
+        revalidateOnFocus: false,
         revalidateOnReconnect: false,
         shouldRetryOnError: false,
-        dedupingInterval: 300000, 
-        revalidateOnMount: false, 
-        keepPreviousData: true,
+        dedupingInterval: 300000,
         focusThrottleInterval: 300000,
+        keepPreviousData: false,
       }}
     >
       {children}
